Moved around the unit tests to be in a single file for common
